Thêm lấy danh sách User và block Tài Khoản

+++ back-end/src/app/services/UserService.js
@@ -2,15 +2,14 @@ const db = require('../../configs/DbPoolConfig');
 const MangaService = require('./MangaService');
 const convertKeysToCamelCase = require('../../utils/CamelCaseUtil');
 const HandleCode = require('../../utils/HandleCode');
+const { users } = require('../../models/init-models')(require('../../configs/DbConfig'));
+const { Op } = require('sequelize');
 class UserService {
   //#region check like-follow
   async isLikeFollowManga(mangaId, userId, type = 'like') {
     try {
       const table = type === 'like' ? 'favorites' : 'following';
-      const [rows] = await db.query(
-        `SELECT * FROM ${table} WHERE mangaId = ? AND userId = ?`,
-        [mangaId, userId]
-      );
+      const [rows] = await db.query(`SELECT * FROM ${table} WHERE mangaId = ? AND userId = ?`, [mangaId, userId]);
       if (rows.length === 0) {
         return { code: HandleCode.NOT_FOUND };
       }
@@ -24,10 +23,7 @@ class UserService {
   async likeFollowManga(mangaId, userId, type = 'like') {
     try {
       const table = type === 'like' ? 'favorites' : 'following';
-      const [rows] = await db.query(
-        `INSERT INTO ${table} (mangaId, userId) VALUES (?, ?)`,
-        [mangaId, userId]
-      );
+      const [rows] = await db.query(`INSERT INTO ${table} (mangaId, userId) VALUES (?, ?)`, [mangaId, userId]);
       await MangaService.updateMangaLikeFollow(mangaId, type);
     } catch (err) {
       if (err.code === 'ER_DUP_ENTRY') {
@@ -45,10 +41,7 @@ class UserService {
   async unLikeFollowManga(mangaId, userId, type = 'like') {
     try {
       const table = type === 'like' ? 'favorites' : 'following';
-      const [rows] = await db.query(
-        `DELETE FROM ${table} WHERE mangaId = ? AND userId = ?`,
-        [mangaId, userId]
-      );
+      const [rows] = await db.query(`DELETE FROM ${table} WHERE mangaId = ? AND userId = ?`, [mangaId, userId]);
       if (rows.affectedRows === 0) {
         return { code: HandleCode.NOT_FOUND };
       }
@@ -131,6 +124,59 @@ class UserService {
     }
   }
   //#endregion
+
+  getUsers = async (page = 1, limit = 1, search_query, is_blocked) => {
+    try {
+      const offset = (page - 1) * limit;
+      const whereClause = {
+        [Op.or]: [{ UserName: { [Op.like]: `%${search_query}%` } }, { Email: { [Op.like]: `%${search_query}%` } }],
+      };
+      console.log('is_blocked', is_blocked);
+      if (is_blocked) {
+        console.log('is_blocked', is_blocked);
+        whereClause.Status = is_blocked == 1 ? HandleCode.ACCOUNT_STATUS_BLOCKED : HandleCode.ACCOUNT_STATUS_ACTIVE;
+      }
+      const { count, rows } = await users.findAndCountAll({
+        where: whereClause,
+        attributes: ['UserId', 'UserName', 'Email', 'Avatar', 'Status'],
+
+        offset,
+        limit,
+      });
+      const transformedRows = rows.map((row) => {
+        const user = row.toJSON();
+        //user.Avatar = user.Avatar ? `${process.env.BASE_URL}/images/${user.Avatar}` : null;
+        return user;
+      });
+      return {
+        totalItems: count,
+        totalPages: Math.ceil(count / limit),
+        currentPage: page,
+        items: convertKeysToCamelCase(transformedRows),
+      };
+    } catch (err) {
+      throw err;
+    }
+  };
+
+  toggleBlockUser = async (userId) => {
+    try {
+      // console.log('userId', userId);
+      const user = await users.findOne({ where: { UserId: userId } });
+      if (!user) {
+        return { code: HandleCode.NOT_FOUND };
+      }
+      const newStatus =
+        user.Status === HandleCode.ACCOUNT_STATUS_ACTIVE
+          ? HandleCode.ACCOUNT_STATUS_BLOCKED
+          : HandleCode.ACCOUNT_STATUS_ACTIVE;
+
+      await users.update({ Status: newStatus }, { where: { UserId: userId } });
+      return { code: HandleCode.SUCCESS, message: 'User status updated successfully.' };
+    } catch (err) {
+      throw err;
+    }
+  };
 }
 
 module.exports = new UserService();
